update pulled down by IDE

+++ ardesp/Arduino/libraries/IRremoteESP8266/examples/TurnOnKelvinatorAC/TurnOnKelvinatorAC.ino
@@ -1,5 +1,6 @@
-
-#include <IRKelvinator.h>
+/* Copyright 2016 David Conran */
+#include <IRsend.h>
+#include <ir_Kelvinator.h>
 
 IRKelvinatorAC kelvir(D1);  // IR led controlled by Pin D1.
 
@@ -17,17 +18,17 @@ void printState() {
   // Display the encoded IR sequence.
   unsigned char* ir_code = kelvir.getRaw();
   Serial.print("IR Code: 0x");
-  for (int i = 0; i < KELVINATOR_STATE_LENGTH; i++)
+  for (uint8_t i = 0; i < KELVINATOR_STATE_LENGTH; i++)
     Serial.printf("%02X", ir_code[i]);
   Serial.println();
 }
 
-void setup(){
+void setup() {
   kelvir.begin();
   Serial.begin(115200);
   delay(200);
 
-  // Set up what we want to send. See IRKelvinator.cpp for all the options.
+  // Set up what we want to send. See ir_Kelvinator.cpp for all the options.
   // Most things default to off.
   Serial.println("Default state of the remote.");
   printState();
